application_tier_flask/requirements.txt
Flask
flask-cors
mysql-connector-python
python-dotenv
boto3
requests